Add isearch support for tcode-use-input-method
Add a new option value `:im` for `tcode-use-isearch`, which
automatically sets `tcode-use-input-method` to `t`, thus enabling
T-Code within isearch.

However, with this setting, the result of postfix-bushu/mazegaki
conversion in isearch is discarded by the caller of the input method.
Add a post-processing phase that updates the isearch state stack with
the conversion result.