Implement prefix-bushu/mazegaki conversion in isearch
When `tcode-use-input-method` is non-`nil`, isearch prepares a
minibuffer for invoking the input method.  However, that minibuffer is
not suitable for prefix-bushu/mazegaki conversion, because it exits
and discards its contents after each character input.  Instead of
using that minibuffer, prepare a dedicated one for the conversion,
which remains active until the entire conversion is completed.