Fix handling of overriding-terminal-local-map
The following bugs, which occur when `tcode-use-input-method` is
non-`nil`, stem from the same cause and are fixed by this commit:

 - After `indent-rigidly`, typing an ordinary character normally exits
   the special mode and inserts that character, so editing can resume
   immediately. The same should happen when inputting a Japanese
   character, but instead an ASCII character appears as if the input
   method were disabled.
 - Inputting a Japanese character also produces an ASCII character
   when a prefix argument such as `C-u` or `M-3` is used.

Both `indent-rigidly` and commands that set prefix arguents (such as
`universal-argument`) call `set-transient-map`, which in turn
activates `overriding-terminal-local-map`.  When
`overriding-terminal-local-map` is active, `tcode-input-method` stops
processing input events entirely and treats them as ASCII characters.

Fix the handling of `overriding-terminal-local-map` in
`tcode-input-method` to match the behavior of `quail-input-method`: it
now stops processing the input event only if the event is bound in
`overriding-terminal-local-map`.

See https://debbugs.gnu.org/cgi/bugreport.cgi?bug=68338 for the
discussion of the same issue in the Quail input method.