Fix undo amalgamation of insertion and deletion
Typing "a BACKSPACE b" and invoking undo once should delete only "a".
However, the undo records for BACKSPACE and "b" are amalgamated, so
both are undone at once.

This bug occurs with the `japanese-2byte-alnum` input method
(`tc-ja-alnum.el`), and with the T-Code input method when
`tcode-use-input-method` is enabled.  Both modify `last-command` to
`self-insert-command` for unknown reasons.  As a result, when "b" is
typed, BACKSPACE is treated as an insertion and amalgamated with the
insertion of "b".

Remove two `setq` forms in `japanese-2byte-alnum`: one that sets
`last-command` and another that sets `last-command-event`, which also
seems unnecessary.  For T-Coded, remove only the `setq` for
`last-command`.  The variable `last-command-event` is still used by
subcommands called within, such as
`tcode-mazegaki-self-insert-or-convert`, which eventually calls
`self-insert-command` and requires `last-command-event` to be set
properly.

The only possible reason I could find for setting `last-command` is
that `tc-comelete.el` refers to it.  However, this seems to be a bug:
the code uses `last-command` to determine the last command executed by
the user, where it should instead use `this-command`.  In any case,
`tc-comelete.el` is currently broken and would require a complete
rework to become usable again, so the impact of this change can be
safely ignored for now.